# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: HRM & Invoice Management System API for CA Firms
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/change-password:
    put:
      operationId: changePassword
      tags: [users]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /engagements:
    get:
      operationId: getEngagements
      tags: [engagements]
      summary: List engagements
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of engagements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EngagementResponse"
    post:
      operationId: createEngagement
      tags: [engagements]
      summary: Create engagement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEngagementRequest"
      responses:
        "201":
          description: Engagement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngagementResponse"
  /engagements/{id}:
    get:
      operationId: getEngagement
      tags: [engagements]
      summary: Get engagement by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Engagement details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngagementResponse"
    put:
      operationId: updateEngagement
      tags: [engagements]
      summary: Update engagement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEngagementRequest"
      responses:
        "200":
          description: Engagement updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngagementResponse"
  /engagements/{id}/assignments:
    get:
      operationId: getEngagementAssignments
      tags: [engagements]
      summary: Get engagement team assignments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EngagementAssignment"
    post:
      operationId: addEngagementAssignment
      tags: [engagements]
      summary: Assign staff to engagement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngagementAssignment"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, hr_admin, finance_officer, manager, employee, partner, trainee]
        employeeId:
          type: integer
          nullable: true
        phone:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        cnic:
          type: string
          nullable: true
        profilePicture:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - status
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [super_admin, hr_admin, finance_officer, manager, employee, partner, trainee]
        phone:
          type: string
        mobile:
          type: string
        cnic:
          type: string
        employeeId:
          type: integer
      required:
        - email
        - name
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, hr_admin, finance_officer, manager, employee, partner, trainee]
        phone:
          type: string
        mobile:
          type: string
        cnic:
          type: string
        status:
          type: string
          enum: [active, inactive]
        employeeId:
          type: integer
        password:
          type: string
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        mobile:
          type: string
        cnic:
          type: string
        profilePicture:
          type: string
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
...
    EngagementResponse:
      type: object
      properties:
        id:
          type: integer
        engagementCode:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [audit, tax, advisory, accounting, compliance, other]
        status:
          type: string
          enum: [planning, execution, review, completed, on_hold, cancelled]
        description:
          type: string
          nullable: true
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        partnerId:
          type: integer
          nullable: true
        partnerName:
          type: string
          nullable: true
        managerId:
          type: integer
          nullable: true
        managerName:
          type: string
          nullable: true
        budget:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        assignmentCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - engagementCode
        - clientId
        - clientName
        - title
        - type
        - status
        - startDate
        - assignmentCount
        - createdAt
    CreateEngagementRequest:
      type: object
      properties:
        clientId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [audit, tax, advisory, accounting, compliance, other]
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        partnerId:
          type: integer
        managerId:
          type: integer
        budget:
          type: string
        notes:
          type: string
      required:
        - clientId
        - title
        - type
        - startDate
    UpdateEngagementRequest:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [audit, tax, advisory, accounting, compliance, other]
        status:
          type: string
          enum: [planning, execution, review, completed, on_hold, cancelled]
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        partnerId:
          type: integer
        managerId:
          type: integer
        budget:
          type: string
        notes:
          type: string
    EngagementAssignment:
      type: object
      properties:
        id:
          type: integer
        engagementId:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        role:
          type: string
        hoursAllocated:
          type: integer
          nullable: true
        hoursWorked:
          type: integer
        notes:
          type: string
          nullable: true
        assignedAt:
          type: string
      required:
        - id
        - engagementId
        - employeeId
        - employeeName
        - role
        - hoursWorked
        - assignedAt
    CreateAssignmentRequest:
      type: object
      properties:
        employeeId:
          type: integer
        role:
          type: string
        hoursAllocated:
          type: integer
        notes:
          type: string
      required:
        - employeeId
        - role
...