# lib/api-zod/src/generated/types/createEngagementRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HRM & Invoice Management System API for CA Firms
 * OpenAPI spec version: 1.0.0
 */
import type { CreateEngagementRequestType } from "./createEngagementRequestType";

export interface CreateEngagementRequest {
  clientId: number;
  title: string;
  type: CreateEngagementRequestType;
  description?: string;
  startDate: string;
  endDate?: string;
  partnerId?: number;
  managerId?: number;
  budget?: string;
  notes?: string;
}